+++ src/CleanRegex/Internal/Match/Details/Groups/DetailGroups.php
@@ -5,7 +5,6 @@ use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
 use TRegx\CleanRegex\Internal\GroupKey\Signatures;
 use TRegx\CleanRegex\Internal\Match\Details\Group\GroupFacade;
 use TRegx\CleanRegex\Internal\Match\Details\Group\GroupHandle;
-use TRegx\CleanRegex\Internal\Match\Details\Group\MatchGroupFactoryStrategy;
 use TRegx\CleanRegex\Internal\Model\Entry;
 use TRegx\CleanRegex\Internal\Model\GroupKeys;
 use TRegx\CleanRegex\Internal\Pcre\Legacy\MatchAllFactory;
@@ -22,7 +21,7 @@ class DetailGroups
     public function __construct(Subject $subject, Signatures $signatures, MatchAllFactory $allFactory, GroupKeys $groupKeys)
     {
         $this->groupKeys = $groupKeys;
-        $this->facade = new GroupFacade($subject, new MatchGroupFactoryStrategy(), $allFactory, new GroupHandle($signatures), $signatures);
+        $this->facade = new GroupFacade($subject, $allFactory, new GroupHandle($signatures), $signatures);
     }
 
     public function groups(GroupArrayKey $arrayKey, UsedForGroup $forGroup, Entry $entry): array