+++ src/CleanRegex/Internal/Pcre/Legacy/UsedForGroup.php
@@ -12,7 +12,6 @@ interface UsedForGroup
     /**
      * @see MatchDetail::get
      * @see MatchedGroup
-     * @see GroupsFacade which is called by everything that calls {@see getGroupTextAndOffset}
      */
     public function isGroupMatched($nameOrIndex): bool;