+++ src/CleanRegex/Replace/Callback/MatchGroupStrategy.php
@@ -2,7 +2,7 @@
 namespace TRegx\CleanRegex\Replace\Callback;
 
 use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
-use TRegx\CleanRegex\Replace\Details\Group\ReplaceGroup;
+use TRegx\CleanRegex\Match\Group;
 use TRegx\CleanRegex\Replace\Details\ReplaceDetail;
 
 class MatchGroupStrategy implements ReplaceCallbackArgumentStrategy
@@ -15,7 +15,7 @@ class MatchGroupStrategy implements ReplaceCallbackArgumentStrategy
         $this->group = $group;
     }
 
-    public function mapArgument(ReplaceDetail $detail): ReplaceGroup
+    public function mapArgument(ReplaceDetail $detail): Group
     {
         return $detail->group($this->group->nameOrIndex());
     }